[autotest] Update deadline for TimedEvent classes
TimedEvent classes have a deadline.  if it's later than that deadline,
ShouldHandle() will return True.  We set that deadline initially when the Event
is constructed, and then we weren't ever updating it.

This Cl introduces UpdateCriteria(), which in TimedEvent classes will move the
deadline forward the appropriate amount of time.

BUG=chromium-os:29848
TEST=unit

Change-Id: Icb54bd6c386d2088d1d6fe2885ddd5c920ae9eed
Reviewed-on: https://gerrit.chromium.org/gerrit/20962
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>