[autotest] Avoid races due to 'version' labels on devices in the lab.
At a high level, this change makes dynamic suites stop deleting the version
labels we use at the end of a run, and also makes the code charged with
creating that label tolerant of discovering that one already exists.

To do the latter, we needed to make the underlying RPC code return
appropriate RPC errors immediately, so that we can detect (and ignore)
them when we try to create a label and it already exists.  Since autotest's
json_rpc just had one type of exception it could throw, we had to add a new
exception type to proxy.py and handle it appropriately.

BUG=chromium-os:29971
TEST=unit
TEST=pre-create a version label and run a suite.  It should run, not error out.
TEST=Run a suite without the label existing.  Should also run.

Change-Id: I71de5670439c6933d1523ae65f7bfb6bc66c4f88
Reviewed-on: https://gerrit.chromium.org/gerrit/21124
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>