[autotest] Fix status timestamps in suite test logs
The dynamic suite code has been logging false timestamps for
tests that it spawns off as part of a suite.  It was timestamping
the beginning and end at the same time.  This pulls timestamp info
from the actual jobs that run the tests and uses those when
creating the log for the full suite.

BUG=chromium-os:30140
TEST=unit
TEST=./site_utils/run_suite.py -b x86-mario -i x86-mario-release/R20-2203.0.0 -s dummy

Change-Id: Ie7d5ababd4b9e9c6c9cdad5718009dad88954598
Reviewed-on: https://gerrit.chromium.org/gerrit/21490
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>