Initial commit of suite_scheduler.ini plus suite V2 changes.
This is an initial attempt at converting a large portion of the
chromeos_test_config over to using the new suite method. There are a few
tests in here that need to be re-evaluated as far as how they are run and how
they fit in the new paradigm. There are notes to that effect in the
suite_scheduler.ini file.

Add in an upstart config for controlling the suit scheduler.

Add in the following V2 test suites:
control.wificell
control.power
control3g
control3g_gobi
control.network_wifi <--we need to figure out how to run this one.
Add enterprise in to the regression suite.
enterprise_enroll requires machine we can auto enroll i.e. the tpm is not owned.
control.power .. we need to figure out where to run this
control.perfalerts - Run platform_BootPerfServer across 3 machines of each board type.
add omx_test to regression.

Create a perf_build suite.
This is for performance tests that should run every build.
control.pagecycler is the first test that is a part of this. There are more to
come.

Deleted suite control files:

The following have been convereted to the new suites appraoch:
control.bvt
control.network_3gk
control.desktopui_omxtest

The following were no longer run even in the previous scheduler:
control.sbstage: Seaboard staging suite that does not run at all.
control.touch: Basis for touch testing. The two tests are still around and this
is not run anywhere.

Fix permissions on files under test_suites.

TEST=suite_enumerator checks.
BUG=chromium-os:30448
Change-Id: I0771cf8c1d0a4760d617ad50bc020c0aea654046
Reviewed-on: https://gerrit.chromium.org/gerrit/21717
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>