hardware_Trackpad: add support to motion criteria in arbitrary directions
In the functionality select-two_fingers_lifted_2d, we allow the user
to move the cursor in arbitrary directions. Modify the code to check
the cursor movement in both x and y directions.

BUG=chromium-os:30456
TEST=Follow the following steps on a chromebook machine:
$ cd /usr/local/autotest/tests/hardware_Trackpad
To run the full regression test which will take about 12 minutes
$ ../../bin/autotest control.regression_full
It should finish normally. The result could be PASS or FAIL, but
not test error.

Change-Id: I402696f0b4f9273f0fd5a2b4de069b548e0a05e9
Reviewed-on: https://gerrit.chromium.org/gerrit/21757
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>