cros_ui_test: Do not overwrite a legitimate resolv.conf
Test to see if resolv.conf is already pointng to our local server
and that a backup file already exists.  This would indicate that
perhaps the last test was interrupted (crash) and that we do not
want to overwrite our backup with this bogus copy.

BUG=chromium-os:18823
TEST=Visual only.  Watching tree while one builder continues to exhibit
this issue.

Change-Id: I7fb85658c519a5ec8295902411e27112cfb7f3f1
Reviewed-on: https://gerrit.chromium.org/gerrit/21802
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>