autotest: Introduce new GraphicsUITest class.
In the medium run all graphics_* and perf tests should derive from
this class. It provides initialization and cleanup functions
monitoring the health of the GPU. It will warn or fail the test if
the GPU state is not satisfactory.

For now we will use this only with WindowManagerGraphicsCapture
until I am satisfied that it works on all boards.

BUG=30287
TEST=Ran on lumpy.

Change-Id: I7ed46a2e06f4ecd819e36c3cd6e3b93e7c8c854d
Reviewed-on: https://gerrit.chromium.org/gerrit/21807
Commit-Ready: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>