Remove platform_ToolchainOptions from the BVT.
platform_ToolchainOptions was removed from suite_Smoke and canaries long
ago, but crept back in to canaries accidentally when we switched from
VMTest to HWTest.

The full explanation for this from Jim Hebert is included below.

---

Speaking as the (security) guy who initially proposed to kick it out of
suite_Smoke: I supported removing it from Smoke because the test is so
prone to false positives that effectively adding any new dep closed the
tree (or rev'ing any existing dep enough that a binary appeared at a
new filename, e.g. libs with version numbers in their filename). I'm
not familiar with the canaries but I assume this is another tree-closing
situation.

The decision to put it INTO Smoke(/canary) in the first place was based
on the incorrect understanding that this was a high-value (it is) AND
low-false-fail (it is not) test. The test for quite a while had a
false-pass bug in it, which was masking the true run rate of
false-fails we have now been plagued by.

The severity of the reaction (tree closure - we want it opened asap!)
and the high probability of false-fail were combining, IMHO, to people
blindly throwing anything and everything onto the ever-growing whitelist
just to get the tree open again, and then never following up to
investigate
what they'd just whitelisted. That's actually worse than not having the
test at all, because it's making no positive impact on the product, and
wasting developer time constantly hacking it.

I'd rather see this test go red (and not close the tree), which gives us
time to actually stop and consider whether this is a false positive or
perhaps the package is explicitly avoiding some of the compiler flags.

---

BUG=chromium-os:28368
TEST=Remote trybot run.

Change-Id: I5b11c43fe2b29977c6f27e2308b528e1c8b9a702
Reviewed-on: https://gerrit.chromium.org/gerrit/22173
Reviewed-by: Jim Hebert <jimhebert@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>