[autotest] Fix deadline handling in TimedEvent.Merge()
Currently, the scheduler checks if it should handle each event,
sleeps, and then re-reads the config and merges newly-created events
with the existing ones.  This entails updating the deadlines of existing
TimedEvents.  Prior to this CL we were doing this unconditionally.

Here's why this is bad:
  When the deadline passes while we're sleeping, the new TimedEvents
created during config re-loading decide that it's too late to run
today, so they set their deadline for the next possible time; e.g. the
new Nightly would set its deadline for tomorrow night.  The scheduler
would then merge that into the existing Nightly event and clobber its
deadline.  _Then_ we check if we should handle the TimedEvents.  Since
the new deadline hasn't passed yet -- because we _just_ changed
it to be tomorrow -- we don't handle the event.

This CL updates the Merge() logic in TimedEvent subclasses to check
whether relevant portions of the deadline have changed, and clobber
the deadline iff they have.  So, Nightly checks to see if the deadline
time of day has changed and merges that in iff it has.  Weekly checks
time of day and day of week.

BUG=chromium-os:30742
TEST=unit
TEST=run the scheduler with Nightly set to run some tasks.  They should get scheduled.
STATUS=Fixed

Change-Id: I859376ea45be5733b65720185f37d6d237da6dc5
Reviewed-on: https://gerrit.chromium.org/gerrit/22286
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>