Force FD_CLOEXEC on the pipe to logging subprocesses.
Logging subprocesses created by LoggingManager are meant to
terminate when the parent process closes its end of a pipe.
The pipe file descriptor is created without setting 'close
on exec'.

If the parent process later uses subprocess.Popen to create a
differnt child process, that child will inherit the open pipe.
The result is that the logging subprocess will fail to see
EOF, fail to terminate, and leave the parent process stuck
waiting indefinitely.

The fix is to force 'close on exec' to be set on the pipe
file descriptor, so that it cannot be inherited by other
subprocesses.

BUG=chromium-os:30806
TEST=run the test included in the bug report

Change-Id: I4fe5ce98d5935fb2ddea143991ef0267033e47e6
Reviewed-on: https://gerrit.chromium.org/gerrit/22429
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>