Update test_scheduler.py to look for r21 and r22.
We white list TOT builds in test_scheduler for now until we
can move to suite_scheduler. R20 was cut recently so R21 builds were not
running.

I extended this to r22 as a buffer but within the next iteration all of TOT
should be running on the new scheduler.

TEST=None
BUG=chromium-os:30844

Change-Id: I8039c533b64d2f54aa81ea3b03273b31c5b5a7f9
Reviewed-on: https://gerrit.chromium.org/gerrit/22541
Reviewed-by: Mike Truty <truty@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>