Extract PyAuto connection creation from ServoTest.
This adds additional methods to SiteHost for creating and managing
a connection to an XMLRPC server on the target host.  Additionally,
other initialization specific to PyAuto connection has moved to a
new server.cros.pyauto_proxy module.

The platform_ServoPyAuto test is updated to use the new framework;
other tests are not yet converted.

BUG=chromium-os:29897
TEST=run platform_ServoPyAuto

Change-Id: If334a1e8db4cddba2b47467c0f747704cbcf34ce
Reviewed-on: https://gerrit.chromium.org/gerrit/22745
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>