hardware_Trackpad: fix the gesture file path for regression test
The board name of cr48 is mario in '/etc/lsb-release'. This CL fixes
a gesture file path problem on mario. The directory is renamed from
cr48 to mario so that the correct gesture files path could be
selected based on the probed board name.

BUG=chromium-os:29516
TEST=Run the short trackpad regression test on mario:
$ cd /usr/local/autotest/tests/hardware_Trackpad
$ ../../bin/autotest control.regression
Should observe:
*** Total pass rate: xx %
*** Total number of (passed / tested) files: (xx / 11)
(The pass rate may or may not be 100% though.)

Change-Id: Ibd81e7decbe9b921d837280f98c0ed810b47efc7
Reviewed-on: https://gerrit.chromium.org/gerrit/22784
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>