hardware_Trackpad: handle boards without gesture sets
Some boards may not be specified gesture sets. The reasons
could be
(1) the gesture files are not captured yet, or
(2) the board does not have a built-in trackpad.

This CL handles the exception of this situation.

BUG=chromium-os:29516
TEST=Run the short trackpad regression test on all models:
It should execute normally. Observe that it does not
incur test error on models without trackpads.
$ cd /usr/local/autotest/tests/hardware_Trackpad
$ ../../bin/autotest control.regression

Change-Id: I59ce2bbcb5e66b43d3bb7aa438c3288b8526cd0f
Reviewed-on: https://gerrit.chromium.org/gerrit/22785
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>