hardware_Trackpad: fix the name change of the result log
The autotest result log was renamed from client.INFO to
client.0.INFO. This patch fixed the problem of the name
change by querying the INFO filename from the logger
handlers.

BUG=chromium-os:29516
TEST=Run the short trackpad regression test any model.
$ cd /usr/local/autotest/tests/hardware_Trackpad
$ ../../bin/autotest control.regression

After execution is finished, chech the result log
/var/tmp/trackpad_test_data/results/user_XXXX
The log should consist of the pass rate, the Verification Log,
and a long list of detailed logs for every gestures.

Change-Id: If3dbb7881ab2e2787a3ea4ced13086950f3ab202
Reviewed-on: https://gerrit.chromium.org/gerrit/22789
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>