[autotest] Extend _KERNEL_UPDATE_TIMEOUT in site_host.py
site_host.ChromiumOSHost implements a method that knows how to
install a new cros image on a device and do some checks to
make sure that it's not obviously bad.  This has a timeout
that it's willing to wait for update_engine to mark the new
kernel as good.  That timeout is currently only 15s longer
than the time that update_engine waits before even trying to
check on the kernel.  We run afoul of this at times.

Extending the timeout costs us no correctness and reduces
flakiness.  There's not a good way to determine reliably
when we should start waiting for this timeout, so this
method is the lesser of known evils.

BUG=chromium-os:30953
TEST=run_suite.py, which will reimage devices as a matter of course

Change-Id: Iec2d85d3412ccebaf8821c7f31ffeb9b8bdbf50e
Reviewed-on: https://gerrit.chromium.org/gerrit/22839
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>