CHROMIUM: autotest: wifi: Test delayed retry of AP
This test run is designed to check that the connection manager
re-scans a few times before quiescing after losing a connection.
We simulate this by shutting off an AP in full sight of the
DUT, waiting for a bit, and then re-instating the AP.  We choose
a hidden AP since this requires the connection manager to do the
scans (since only the connection manager knows to scan for them,
as opposed to wpa_supplicant).

BUG=chromium-os:30984
TEST=This is a test

Change-Id: If2767149630d21daf9233a3ecf2d694497a8c614
Reviewed-on: https://gerrit.chromium.org/gerrit/23231
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>