Add a test for SD card reader that does not depend on fio.
We observed some issues when doing the SD card reader test using fio.
Also, performance test is not necessarily needed since we mainly focus
on the functionality of the card reader in factory test. Thus, we
replace the original test with a new test that does not depend on fio.
We implement the following tests:
   * Read/Write test
   * Lock (write protection) test

BUG=chrome-os-partner:9728
TEST=Checked SD card reader read/write test works.
     Switched the lock on during read/write test, and the test failed.
     Checked SD card lock test works.
     Switched the lock off during lock test, and the test failed.
     Checked the new test integrates with existing factory test.

Change-Id: Icf810a7ef55ec7d9d13cc41b4da73282f6e3d2c1
Reviewed-on: https://gerrit.chromium.org/gerrit/23487
Reviewed-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Ready: Ricky Liang <jcliang@google.com>
Tested-by: Ricky Liang <jcliang@google.com>