factory: Add ALS calib. test and improve the camera perf test
The CL adds the ALS calibration feature for the Camera/ALS test.
The test will now control the test fixture through RS232 serial
port and record ALS values for different light settings. The CL
also refactors the i2c library to expose useful functions.

Besides, the CL improves the camera performance test in several
aspects. 1) The test now captures multiple images and average
them to reduce the image noise. 2) The test now supports an
'unit-test' mode so that the developers can verify the test
integrity when the test fixture is not available. 3) The test
now computes the image shift and tilt amount and check if they
are within the spec. Finally, the CL adds a sample test
parameter file under the src directory so that the developer can
have a reference.

BUG=chrome-os-partner:6616,chrome-os-partner:6617,chrome-os-partner:9992
TEST=Test it on link and on the test fixture -> works

Change-Id: Ic79c6cc9bef0e0e3c5c3833b15ab4eac6720a044
Reviewed-on: https://gerrit.chromium.org/gerrit/23501
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>
Commit-Ready: Tai-Hsu Lin <sheckylin@chromium.org>