Check main disk is fixed in hardware_SsdDetection
This CL checks /sys/block/[DEVICE]/removable to make sure that
main disk is fixed in hardware_SsdDetection test.
This is required by partner that we should test more properties of eMMC.

BUG=None
TEST=run hardware_SsdDetection on DUT. I have tested it on snow board and
lumpy and can pass the test.
We plan to test other properties of eMMC in the future.

Change-Id: Id1f09caaf50f4649b03f95972ce5a383597c37f1
Reviewed-on: https://gerrit.chromium.org/gerrit/24011
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>