[autotest] Only remove labels that are actually on a host
Before reimaging a device in a dynamic_suite run, we clear off
any labels on the device that my purport to indicate what version
is currently installed.  We were doing this by enumerating all
version labels known to autotest and then blanket-removing them all
from the device whether they were actually on it or not.  This was
simple and, when we only had a few dozen labels in play at a time,
fast enough.  Now that we have ~1500, it's awful.  So, enumerate
only the version labels that are actually on a host, and remove
only those instead.

BUG=chromium-os:31381
TEST=Run a suite, and watch your autotest rpcserver.log to see that only one remove_labels() RPC comes in per host being reimaged.
STATUS=Fixed

Change-Id: I59503e4bcc72b99a33c85f018917486b3aafcf42
Reviewed-on: https://gerrit.chromium.org/gerrit/24052
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>