[autotest] Tool to gather suite dependencies at build-time
In order to support DEPENDENCIES in dynamic suites, the infrastructure needs
to be able to determine the needs of a suite _before_ doing any reimaging work.
Naively, this would require looking at all controls files in a suite before
picking machines to reimage.  Staging all control files on the dev server is
slow, and we currently do this in the background while reimaging devices.

To avoid forcing the infrastructure to stage the entire autotest tarball before
doing any work, we plan to pre-process control files at build time and use
this data to make decisions at run-time.  This tool does the pre-processing
work.

BUG=chromium-os:22060
TEST=run this tool, view output.

Change-Id: Ifde836db1a60eff2e7fb5eb74ba4ae84bd505d5f
Reviewed-on: https://gerrit.chromium.org/gerrit/24217
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>