Introduce a clean_staged_images script for pruning the devserver and labels.
clean_staged_images is responsible for removing X amount of "old" builds on
on the chrome os devserver as well as pruning any labels of type
cros-version:... from the Autotest DB of any build that is no longer present
on the devserver.

clean_staged_images is somewhat devserver centric but also needs to interact
with autotest to clean up after suite runs that created labels. As a result
this has landed in autotest as opposed to platform/dev.

BUG=chromium-os:31260
TEST=Local runs on experimental data from the old devserver.

Change-Id: Ie23c647a2c390aea8e8ba7ec6b0f6769053420fd
Reviewed-on: https://gerrit.chromium.org/gerrit/24309
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>