Restore the 'perfalerts' suite control file to working order.
The test included code to that that the suite would only run
immediately after a clean install, and to install again if the test
failed.  Unfortunately, the install code was unable to succeed
because no image name was being passed to install from.  The test
is now largely unnecessary because with Suites v2, the test will
always run from (nearly) immediately after the re-install.

BUG=chromium-os:31149
TEST=run_remote_tests

Change-Id: Ie4cd21616c6139b81f416c85834cd32b8c418604
Reviewed-on: https://gerrit.chromium.org/gerrit/24901
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Mike Truty <truty@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>