3G: Wait for modem to go into disabled state before enabling
When autotest resets a modem, it disables the modem and then enables the
modem.  There is a short 1 second sleep betweeen disable and enable.
Some times, the modem takes longer than 1 second to disable so the
enable fails because the modem is still going down.  This CL makes sure the
modem goes to a disabled state first before enabling.

BUG=chromium-os:31801
TEST=cellular_Smoke

Change-Id: I6afc2078d192eed6a47d72abd4afeaab9fca9a23
Reviewed-on: https://gerrit.chromium.org/gerrit/25157
Reviewed-by: Gary Morain <gmorain@chromium.org>
Commit-Ready: Thieu Le <thieule@chromium.org>
Tested-by: Thieu Le <thieule@chromium.org>