Autotest: Clean up gs_offloader.
Changed the structure of gs_offloader so that it will now loop
continuously over the results directly.

Added logging to a file and replaced all print out's with logging
messages.

Added a timeout to shelling out commands. Set the timeout to three
hours and utilized signal and SIGALRM so that the offloader command will
be terminated once the timeout expires.

BUG=chromium-os:29571
TEST=Ran the script (without folder deletion and is_job_complete) on
chromeos-autotest.cbf

Change-Id: I97b47db93d2f5e373dff9c9fa037f13f38f146cb
Reviewed-on: https://gerrit.chromium.org/gerrit/26370
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Simran Basi <sbasi@google.com>
Tested-by: Simran Basi <sbasi@google.com>