Add in random server fixes for the move to the new Atlantis machines.
Update settings.py to contain a SESSION_COOKIE_AGE of 1200 seconds. this is a
temporary hack to keep the number of django sessions down via the autotest
scheduler clean up process. Every new RPC seems to be creating a new
django_session entry and this needs to be futher evaluated.

Update global_config.ini to enable atlantis2 as a drone. This moves the
autoserv processes off of the main database server and opens the door for
further horizontal scaling as we need it later on.

Set the process limit for atlantis2 to 600 so we do not kill the machine.

Add in a dummy site_archive_results so that we do not copy results back to
atlantis1 AKA cautotest. This copy is unnecessary since we move all of our
results to Google Storage via gs_offloader.

TEST=Running live on the server now.
BUG=None

Change-Id: I77fc08c543092c21acd0cc238ea04c91527c9738
Reviewed-on: https://gerrit.chromium.org/gerrit/27057
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>