Update test_scheduler.
Change all configs to point to the shorter platform name in the json config
files.

Update test_scheduler to always add a pool:suites to the job being created so
that we do not use machines with the same platform label in a different pool.
(Like the bvt machines).

TEST=validate_json_files and test_scheduler --verbose run with job creation/
validation
BUG=chromium-os:32449

Change-Id: I6f9ec2dd72741c2f3f2dd9ecbd9b366d7a76b45f
Reviewed-on: https://gerrit.chromium.org/gerrit/27092
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>