[autotest] Refine bgscan test methodology
Change from subscribing to mac80211 event summaries to measuring ping
latencies as a proxy for whether or not background scans have impacted
our foreground traffic performance.

At some point, we decided we should abort mac80211 background scans when
we see data in the transmit buffer.  We tested that we were in fact
doing this by subscribing to mac80211 events on the client, and looking
for messages that background scans had started and been aborted.  This
was a problem because some chips handle this in hardware, and bypass our
driver with its patch to abort bgscans.  By measuring traffic latency
rather than than looking for events, we get closer to testing impact to
the user, rather than some strange strings in a log somewhere.

We do this by measuring ping latency with and without background scans
enabled and comparing the round trip times.  If the times are similar
enough, then we are satisfied that bgscan isn't
a problem.

TEST=Ran test w/ and w/o driver patch, observed appropriate failures
BUG=chromium-os:31435

Change-Id: I670ad285176fdcd642f63071d39bccb5074d44a8
Reviewed-on: https://gerrit.chromium.org/gerrit/27213
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Christopher Wiley <wiley@chromium.org>