power_Resume: return suspend/resume time for devices
Return as keyvals the suspend and resume time for the slowest n devices.
The keys have the following format: seconds_dev_<devname>_<action>.
Where <devname> is the name of the device and <action> is one of the
following strings: 'resume' or 'suspend'. For each device, the suspend
and resume time is logged to the debug log.

BUG=chromium-os:31949
TEST=works on a Samsung Series 5 550 machine

Change-Id: I481df687363850084ab7c4b916247d0156ecb071
Signed-off-by: Lucian Cojocar <cojocar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27207
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>