Add BVT for Lucas/Daisy to suite_scheduler.
I want to let this bake for a little while before we decide whether
or not we should put this in the HWTest step. In general we should
adopt this as a standard way of proving a new board is ready to go
in to the waterfall.

BUG=None
TEST=None

Change-Id: I1d0c8f85b45a6c943c74cef067dd6c4a5a9e6e50
Reviewed-on: https://gerrit.chromium.org/gerrit/27230
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>