Add in test_suites/control.power_requirements.
While we'd added power_weekly to the suite scheduler, we didn't actually
have a suite control file for it.  This CL renames power_weekly to be called
something that better reflects the intended purpose of the tests in the suite
and schedules it to be run weekly.

BUG=None
TEST=None

Change-Id: I333ffd5b829a9b198d3f1a0cbf5803ce47abf460
Reviewed-on: https://gerrit.chromium.org/gerrit/27232
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>