Change _WARNING_DURATION in drone manager from 60 to 400.
Until we have a handle on what we expect these servers to be able to
process I am increasing this time over the current threshold of warnings
we are seeing.

This timeout only indicates that things are processing slowly and nothing
dire. This information can be pulled from the logs when we prioritize this
side of our server setup.

BUG=None
TEST=Syntax

Change-Id: Ia205d03a78f279df8eb72dd2e04c5e5ee4598b59
Reviewed-on: https://gerrit.chromium.org/gerrit/27451
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>