Add functionality to disable collection of core files through the config.
We add CLIENT->collect_corefiles which is not set is always set to True.
This must be in CLIENT as it's the only stanza of the config passed to
clients and sysinfo is collected on the target not from the server.

BUG=chromium-os:32725
TEST=Ran platform_BootPerfServer (with 1 iteration) and bad build that was
crashy.  Toggled config both ways and saw us filter core files when
necessary

Change-Id: I69ed34eb7661ebe63eb5b9c578bcd1394f2298ae
Reviewed-on: https://gerrit.chromium.org/gerrit/27831
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Yu-Ju Hong <yjhong@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>