[autotest] Continue polling for hosts when waiting for a reimage to start running
If the lab is loaded, a reimaging job that's scheduled across multiple
DUTs can sometimes have to wait a while in order for all of it's
autoupdate 'tests' to get hosts assigned.  The trick is that the
hostless reimaging job will be listed as 'Running' even before
it has all its hosts assigned.  job_status.wait_for_and_lock_job_hosts()
did not deal correctly with this case, and so it would not handle
hosts that appeared after it initially queries the reimaging job
for the names of the DUTs it's using.

Instead, every time we're about to see if we've discovered and locked
all hosts, we poll to check if new hosts have been added to the job.
In order to make sure we don't bail early, job_status.gather_job_hostnames()
returns placeholders for tests that don't yet have hosts assigned.

BUG=chromium-os:30978
TEST=unit
TEST=run_suite while holding at least one required host out of the pool until after reimaging has begun.  Then re-add it to the pool and check to see that both hosts get locked while reimaging, and stay locked until after tests have been scheduled.

Change-Id: I7937313e32ddb4b3fb221fdda0e03a210ad8d71b
Reviewed-on: https://gerrit.chromium.org/gerrit/27833
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>