For recovery, use cold reset to force the DUT to be off.
When booting in recovery, power off the unit using cold reset,
so that we don't need to know the power state of the DUT, or
require the DUT to have working software installed.

This change will allow for robust automated recovery when a test
image bricks units in the lab.

BUG=chromium-os:32714
TEST=platform_InstallTestImage from a variety of DUT states

Change-Id: I09db465cfe34486a42132efaeee71997d5a750ac
Reviewed-on: https://gerrit.chromium.org/gerrit/27840
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>