autotest: Fix flakiness when connecting to WiFi network.
Occasionally, connecting to a WiFi service would fail because
wpa_select_time in site_wlan_connect.py reported a negative value.
wpa_select_time is supposed to be the time between finding a service
in the scan results and authenticating to it.  Those two events
(finding a service and authenticating) are reported by two different
processes (conenction manager (flimflam/shill) and wpa_supplicant).
When the times for the two events are close, it is possible that
wpa_supplciant reports the authentication before connection manager
reports finding the service, resulting in a negative value in
wpa_select_time.

The fix is simply to set wpa_select_time to zero when it is negative.

BUG=chromium-os:32671
TEST=Ran network_WiFiRoaming.001SSIDSwitchBack

Change-Id: I35102041172bd6c830d68f988a028fdffb5b9f32
Reviewed-on: https://gerrit.chromium.org/gerrit/27900
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Gary Morain <gmorain@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>